Rental Edit & Delete Fixed UI Updates

+++ Controllers/RentalController.cs
@@ -5,7 +5,7 @@ using Microsoft.EntityFrameworkCore;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
-
+using Microsoft.AspNetCore.Mvc.Rendering;
 
 public class RentalController : Controller
 {
@@ -56,24 +56,55 @@ public class RentalController : Controller
         {
             return NotFound();
         }
+        ViewBag.AvailabilityOptions = new SelectList(new List<string> { "true", "false" });
+        ViewBag.StatusOptions = new SelectList(new List<string> { "Available", "Rented", "Maintenance" });
         return View(rental);
     }
 
-    // POST: Rental/Edit/5
+    // POST: Rental/Edit/
     [HttpPost]
     [ValidateAntiForgeryToken]
-    public async Task<IActionResult> Edit(int id, Rental rental)
+    public async Task<IActionResult> Edit(int RentalId, [Bind("RentalId, VehicleName, VehicleType, State, Country, RentalCost, Availability, Status, LicensePlate, MakeModel, Description, PickupLocation, DropoffLocation")]
+    Rental rental)
     {
-        if (id != rental.RentalId)
+        if (RentalId != rental.RentalId)
         {
             return NotFound();
+        }
+
+        if (!ModelState.IsValid)
+        {
+            var errors = ModelState.Values.SelectMany(v => v.Errors);
+            foreach (var error in errors)
+            {
+                // Log or inspect the error message
+                Console.WriteLine(error.ErrorMessage);
+            }
         }
 
         if (ModelState.IsValid)
         {
             try
             {
-                _context.Update(rental);
+                var rentalToUpdate = await _context.Rentals.FindAsync(RentalId);
+                if (rentalToUpdate == null)
+                {
+                    return NotFound();
+                }
+
+                rentalToUpdate.VehicleName = rental.VehicleName;
+                rentalToUpdate.VehicleType = rental.VehicleType;
+                rentalToUpdate.State = rental.State;
+                rentalToUpdate.Country = rental.Country;
+                rentalToUpdate.RentalCost = rental.RentalCost;
+                rentalToUpdate.Availability = rental.Availability;
+                rentalToUpdate.Status = rental.Status;
+                rentalToUpdate.LicensePlate = rental.LicensePlate;
+                rentalToUpdate.MakeModel = rental.MakeModel;
+                rentalToUpdate.Description = rental.Description;
+                rentalToUpdate.PickupLocation = rental.PickupLocation;
+                rentalToUpdate.DropoffLocation = rental.DropoffLocation;
+
                 await _context.SaveChangesAsync();
             }
             catch (DbUpdateConcurrencyException)
@@ -101,7 +132,7 @@ public class RentalController : Controller
         }
 
         var rental = await _context.Rentals
-            .FirstOrDefaultAsync(m => m.RentalId == id);
+            .FirstOrDefaultAsync(r => r.RentalId == id);
         if (rental == null)
         {
             return NotFound();
@@ -111,14 +142,19 @@ public class RentalController : Controller
     }
 
     // POST: Rental/Delete/5
-    [HttpPost, ActionName("Delete")]
+    [HttpPost, ActionName("DeleteConfirmed")]
     [ValidateAntiForgeryToken]
-    public async Task<IActionResult> DeleteConfirmed(int id)
+    public async Task<IActionResult> DeleteConfirmed(int RentalId)
     {
-        var rental = await _context.Rentals.FindAsync(id);
-        _context.Rentals.Remove(rental);
-        await _context.SaveChangesAsync();
-        return RedirectToAction(nameof(Index));
+        var rental = await _context.Rentals.FindAsync(RentalId);
+        if (rental != null)
+        {
+            _context.Rentals.Remove(rental);
+            await _context.SaveChangesAsync();
+            return RedirectToAction("Index");
+        }
+
+        return NotFound();
     }
 
     private bool RentalExists(int id)
